requirements.txt
fastapi
uvicorn
websockets
python-multipart
ultralytics
paddleocr
paddlepaddle
opencv-python
numpy
pillow
python-dotenv
pydantic
requests
huggingface-hub
roboflow
